python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "easyino"
version = "0.1.0"
authors = [
  { name="Rafael Geovane", email="rafawloficial@gmail.com" },
]
description = "Extremely simple Python to Arduino bridge library."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial>=3.5",
]

[project.urls]
"Homepage" = "https://github.com/Faelzin09663/easyINO"

[tool.setuptools.packages.find]
include = ["easyino*"]
